adding more information options

+++ app.py
@@ -186,16 +186,18 @@ def user_cart_delete():
 @app.route('/<user_id>/delete')
 def user_delete(user_id):
     _user = users.find_one({'_id': ObjectId(user_id)})
-    user_inventory = inventories.find_one({'_id': ObjectId(_user['inventory'])})
-    user_items = items.find({'inventory': ObjectId(_user['inventory'])})
+    if _user['inventory'] is not None:
+        user_inventory = inventories.find_one({'_id': ObjectId(_user['inventory'])})
+        user_items = items.find({'inventory': ObjectId(_user['inventory'])})
     cart_items = carts.find({'user_id': ObjectId(_user['_id'])})
     for item in cart_items:
         carts.delete_one({'_id': ObjectId(item['_id'])})
-    for item in user_items:
-        item_in_other_cart = carts.find({'item_id': ObjectId(item['_id'])})
-        for _item in item_in_other_cart:
-            carts.delete_one({'_id': ObjectId(_item['_id'])})
-        items.delete_one({'_id': ObjectId(item['_id'])})
+    if _user['inventory'] is not None:
+        for item in user_items:
+            item_in_other_cart = carts.find({'item_id': ObjectId(item['_id'])})
+            for _item in item_in_other_cart:
+                carts.delete_one({'_id': ObjectId(_item['_id'])})
+                items.delete_one({'_id': ObjectId(item['_id'])})
     if _user['inventory'] is not None:
         inventories.delete_one({'_id': ObjectId(user_inventory['_id'])})
     users.delete_one({'_id': ObjectId(user_id)})
